debug the updated the thoughts
