--- src/components/ListMessage.jsx
@@ -1,3 +1,0 @@
-export const ListMessage = () => {
-  return <div></div>;
-};
