+++ src/components/PostMessage.jsx
@@ -1,8 +1,22 @@
 import { useState, useEffect } from "react";
-export const PostMessage = ({ fetchPosts }) => {
+import { MessageList } from "./MessageList";
+export const PostMessage = () => {
   //Initialize the state is empty string
   const [newPost, setNewPost] = useState("");
   const [error, setError] = useState(""); //initial state for error message
+  const [thoughts, setThoughts] = useState([]);
+  const apiUrl = "https://happy-thoughts-ux7hkzgmwa-uc.a.run.app/thoughts";
+  const fetchPosts = () => {
+    // Fetch recent thoughts, this will return the latest 20 thoughts from API
+    fetch(apiUrl)
+      .then((response) => response.json())
+      .then((json) => setThoughts(json))
+      .catch((err) => console.log(err))
+      .finally(() => {
+        setLoading(false);
+      });
+  };
+
   useEffect(() => {
     if (newPost.length >= 141) {
       setError("Your message is too long, please reset your message😞");
@@ -10,11 +24,13 @@ export const PostMessage = ({ fetchPosts }) => {
       setError("");
     }
   }, [newPost]);
+
   const handleFormSubmit = (event) => {
     event.preventDefault();
-    console.log("New Post❤️", newPost);
+    // console.log("New Post❤️", newPost);
     if (newPost.length <= 4) {
       setError("Your message is too short, it needs at least 5 letters😞");
+    } else {
       const options = {
         method: "POST",
         body: JSON.stringify({
@@ -26,49 +42,55 @@ export const PostMessage = ({ fetchPosts }) => {
       fetch("https://happy-thoughts-ux7hkzgmwa-uc.a.run.app/thoughts", options)
         .then((res) => res.json())
         .then((json) => {
-          fetchPosts(json);
+          setThoughts((prevThoughts) => [json, ...prevThoughts]);
+          // setNewPost("");
         })
+        .catch((err) => console.log(err))
         .finally(() => {
           setNewPost("");
         });
     }
   };
+
   return (
     <>
-      {/* Form part */}
-
-      <form onSubmit={handleFormSubmit}>
-        <h2>What is making you happy right now?</h2>
-        {/* Text part  */}
-        <textarea
-          rows="3"
-          placeholder="The feeling of accomplishment after completing a challenging task or achieving a long-sought goal."
-          value={newPost}
-          onChange={(e) => setNewPost(e.target.value)}
-        ></textarea>
-        {/* Control Post-length and Error Message */}
-        <div className="post-length">
-          <p className="error">{error}</p>
-
-          <p className={`length ${newPost.length >= 140 ? "red" : ""}`}>
-            {newPost.length}/140
-          </p>
-        </div>
-        {/* Submit thoughts Button */}
-        <button
-          type="submit"
-          id="submitPostBtn"
-          aria-label="button for submitting your post"
-        >
-          <span className="emoji" aria-label="heart emoji">
-            ❤️
-          </span>
-          Send Happy Thoughts
-          <span className="emoji" aria-label="heart emoji">
-            ❤️
-          </span>
-        </button>
-      </form>
+      <div className="post-wrapper">
+        {/* Form part */}
+        <form onSubmit={handleFormSubmit}>
+          <h2>What is making you happy right now?</h2>
+          {/* Text part  */}
+          <textarea
+            rows="3"
+            placeholder="The feeling of accomplishment after completing a challenging task or achieving a long-sought goal."
+            value={newPost}
+            onChange={(e) => setNewPost(e.target.value)}
+          ></textarea>
+          {/* Control Post-length and Error Message */}
+          <div className="post-length">
+            <p className="error">{error}</p>
+            <p className={`length ${newPost.length >= 140 ? "red" : ""}`}>
+              {newPost.length}/140
+            </p>
+          </div>
+          {/* Submit thoughts Button */}
+          <button
+            type="submit"
+            id="submitPostBtn"
+            aria-label="button for submitting your post"
+          >
+            <span className="emoji" aria-label="heart emoji">
+              ❤️
+            </span>
+            Send Happy Thoughts
+            <span className="emoji" aria-label="heart emoji">
+              ❤️
+            </span>
+          </button>
+        </form>
+      </div>
+      <div className="list-wrapper">
+        <MessageList thoughts={thoughts} newPost={newPost} />
+      </div>
     </>
   );
 };
